swift-chart-gapgesture-unfin-skipped

+++ Swift/IOS16/Projects/007.Charts/Chart Demo 2/Chart Demo 2/Data File.swift
@@ -4,7 +4,7 @@ import SwiftUI
 struct DailySalesType: Identifiable {
     let id = UUID()
     let day: String
-    let sales: Int
+    let sales: Double
 }
